Time: 5 ms (5.74%) | Memory: 6.8 MB (41.15%) - LeetSync

+++ leetcode/1746-largest-substring-between-two-equal-characters/largest-substring-between-two-equal-characters.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    int maxLengthBetweenEqualCharacters(string s) {
+    size_t maxLengthBetweenEqualCharacters(string s) {
         unordered_map<char, int> mapping_chars{};
         int max_length{-1};
